Changed functions to arrow functions + moved types
