+++ src/Components/Drawer/domain.tsx
@@ -1,0 +1,5 @@
+import { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
+
+export interface AppBarProps extends MuiAppBarProps {
+    open?: boolean;
+  }
