+++ src/Components/Header/domain.tsx
@@ -1,0 +1,3 @@
+export interface headerProps {
+    toogleOpen: () => void;
+}
